source/requirements.txt
sentence-transformers==2.2.2
huggingface_hub==0.11.1
faiss-cpu==1.7.3
pandas==1.5.3
numpy==1.24.3
torch==2.0.1
requests==2.28.2
google-generativeai
python-dotenv